separate did doc public API and internal types
